pipelines/scripts/requirements.txt
atlassian-python-api==3.28.1
BeautifulSoup4==4.11.1
pandas==1.3.4
lxml==4.6.4
Jinja2==3.0.0